+++ packages/pigeon/platform_tests/test_plugin/ios/Classes/TestPlugin.swift
@@ -28,6 +28,9 @@ public class TestPlugin: NSObject, FlutterPlugin, HostIntegrationCoreApi {
       binaryMessenger: binaryMessenger, messageChannelSuffix: "suffixOne")
     flutterSmallApiTwo = FlutterSmallApi(
       binaryMessenger: binaryMessenger, messageChannelSuffix: "suffixTwo")
+
+    StreamIntsStreamHandler.register(with: binaryMessenger, wrapper: SendInts())
+    StreamEventsStreamHandler.register(with: binaryMessenger, wrapper: SendEvents())
     proxyApiRegistrar = ProxyApiTestsPigeonProxyApiRegistrar(
       binaryMessenger: binaryMessenger, apiDelegate: ProxyApiDelegate())
     proxyApiRegistrar!.setUp()
@@ -1214,6 +1217,53 @@ public class TestPluginWithSuffix: HostSmallApi {
 
 }
 
+class SendInts: StreamIntsStreamHandler {
+  var timerActive = false
+
+  override func onListen(withArguments arguments: Any?, sink: PigeonEventSink<Int64>) {
+    var count: Int64 = 0
+    if !timerActive {
+      timerActive = true
+      Timer.scheduledTimer(withTimeInterval: 0.01, repeats: true) { _ in
+        sink.success(count)
+        count += 1
+        if count >= 5 {
+          sink.endOfStream()
+        }
+      }
+    }
+  }
+}
+
+class SendEvents: StreamEventsStreamHandler {
+  var timerActive = false
+  var eventList: [EventChannelDataBase] =
+    [
+      IntEvent(value: 1),
+      StringEvent(value: "string"),
+      BoolEvent(value: false),
+      DoubleEvent(value: 3.14),
+      ObjectsEvent(value: true),
+      EnumEvent(value: EventEnum.fortyTwo),
+      ClassEvent(value: EventAllNullableTypes(aNullableInt: 0)),
+    ]
+
+  override func onListen(withArguments arguments: Any?, sink: PigeonEventSink<EventChannelDataBase>)
+  {
+    var count = 0
+    if !timerActive {
+      timerActive = true
+      Timer.scheduledTimer(withTimeInterval: 0.01, repeats: true) { _ in
+        if count >= self.eventList.count {
+          sink.endOfStream()
+        } else {
+          sink.success(self.eventList[count])
+          count += 1
+        }
+      }
+    }
+  }
+}
 class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
   func pigeonApiProxyApiTestClass(_ registrar: ProxyApiTestsPigeonProxyApiRegistrar)
     -> PigeonApiProxyApiTestClass
@@ -1221,7 +1271,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
     class ProxyApiTestClassDelegate: PigeonApiDelegateProxyApiTestClass {
       func pigeonDefaultConstructor(
         pigeonApi: PigeonApiProxyApiTestClass, aBool: Bool, anInt: Int64, aDouble: Double,
-        aString: String, aUint8List: FlutterStandardTypedData, aList: [Any?], aMap: [String?: Any?],
+        aString: String, aUint8List: FlutterStandardTypedData, aList: [Any?],
+        aMap: [String?: Any?],
         anEnum: ProxyApiTestEnum, aProxyApi: ProxyApiSuperClass, aNullableBool: Bool?,
         aNullableInt: Int64?, aNullableDouble: Double?, aNullableString: String?,
         aNullableUint8List: FlutterStandardTypedData?, aNullableList: [Any?]?,
@@ -1238,35 +1289,43 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return ProxyApiTestClass()
       }
 
-      func attachedField(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func attachedField(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> ProxyApiSuperClass
       {
         return ProxyApiSuperClass()
       }
 
-      func staticAttachedField(pigeonApi: PigeonApiProxyApiTestClass) throws -> ProxyApiSuperClass {
+      func staticAttachedField(pigeonApi: PigeonApiProxyApiTestClass) throws
+        -> ProxyApiSuperClass
+      {
         return ProxyApiSuperClass()
       }
 
-      func aBool(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func aBool(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> Bool
       {
         return true
       }
 
-      func anInt(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func anInt(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> Int64
       {
         return 0
       }
 
-      func aDouble(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func aDouble(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> Double
       {
         return 0.0
       }
 
-      func aString(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func aString(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> String
       {
         return ""
@@ -1278,7 +1337,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return FlutterStandardTypedData(bytes: Data())
       }
 
-      func aList(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func aList(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> [Any?]
       {
         return []
@@ -1290,7 +1350,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return [:]
       }
 
-      func anEnum(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass) throws
+      func anEnum(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+        throws
         -> ProxyApiTestEnum
       {
         return ProxyApiTestEnum.one
@@ -1302,25 +1363,33 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return ProxyApiSuperClass()
       }
 
-      func aNullableBool(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableBool(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> Bool?
       {
         return nil
       }
 
-      func aNullableInt(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableInt(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> Int64?
       {
         return nil
       }
 
-      func aNullableDouble(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableDouble(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> Double?
       {
         return nil
       }
 
-      func aNullableString(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableString(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> String?
       {
         return nil
@@ -1332,19 +1401,25 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return nil
       }
 
-      func aNullableList(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableList(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> [Any?]?
       {
         return nil
       }
 
-      func aNullableMap(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableMap(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> [String?: Any?]?
       {
         return nil
       }
 
-      func aNullableEnum(pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass)
+      func aNullableEnum(
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass
+      )
         throws -> ProxyApiTestEnum?
       {
         return nil
@@ -1384,7 +1459,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoDouble(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aDouble: Double
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aDouble: Double
       ) throws -> Double {
         return aDouble
       }
@@ -1396,7 +1472,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String
       ) throws -> String {
         return aString
       }
@@ -1540,7 +1617,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoAsyncDouble(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aDouble: Double,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aDouble: Double,
         completion: @escaping (Result<Double, Error>) -> Void
       ) {
         completion(.success(aDouble))
@@ -1554,7 +1632,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoAsyncString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String,
         completion: @escaping (Result<String, Error>) -> Void
       ) {
         completion(.success(aString))
@@ -1591,7 +1670,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
 
       func echoAsyncEnum(
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
-        anEnum: ProxyApiTestEnum, completion: @escaping (Result<ProxyApiTestEnum, Error>) -> Void
+        anEnum: ProxyApiTestEnum,
+        completion: @escaping (Result<ProxyApiTestEnum, Error>) -> Void
       ) {
         completion(.success(anEnum))
       }
@@ -1628,7 +1708,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoAsyncNullableDouble(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aDouble: Double?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aDouble: Double?,
         completion: @escaping (Result<Double?, Error>) -> Void
       ) {
         completion(.success(aDouble))
@@ -1642,7 +1723,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoAsyncNullableString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String?,
         completion: @escaping (Result<String?, Error>) -> Void
       ) {
         completion(.success(aString))
@@ -1657,14 +1739,16 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func echoAsyncNullableObject(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, anObject: Any?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        anObject: Any?,
         completion: @escaping (Result<Any?, Error>) -> Void
       ) {
         completion(.success(anObject))
       }
 
       func echoAsyncNullableList(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aList: [Any?]?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aList: [Any?]?,
         completion: @escaping (Result<[Any?]?, Error>) -> Void
       ) {
         completion(.success(aList))
@@ -1679,7 +1763,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
 
       func echoAsyncNullableEnum(
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
-        anEnum: ProxyApiTestEnum?, completion: @escaping (Result<ProxyApiTestEnum?, Error>) -> Void
+        anEnum: ProxyApiTestEnum?,
+        completion: @escaping (Result<ProxyApiTestEnum?, Error>) -> Void
       ) {
         completion(.success(anEnum))
       }
@@ -1688,13 +1773,15 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
 
       }
 
-      func echoStaticString(pigeonApi: PigeonApiProxyApiTestClass, aString: String) throws -> String
+      func echoStaticString(pigeonApi: PigeonApiProxyApiTestClass, aString: String) throws
+        -> String
       {
         return aString
       }
 
       func staticAsyncNoop(
-        pigeonApi: PigeonApiProxyApiTestClass, completion: @escaping (Result<Void, Error>) -> Void
+        pigeonApi: PigeonApiProxyApiTestClass,
+        completion: @escaping (Result<Void, Error>) -> Void
       ) {
         completion(.success(Void()))
       }
@@ -1770,10 +1857,12 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoDouble(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aDouble: Double,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aDouble: Double,
         completion: @escaping (Result<Double, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoDouble(pigeonInstance: pigeonInstance, aDouble: aDouble) { response in
+        pigeonApi.flutterEchoDouble(pigeonInstance: pigeonInstance, aDouble: aDouble) {
+          response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -1784,10 +1873,12 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String,
         completion: @escaping (Result<String, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoString(pigeonInstance: pigeonInstance, aString: aString) { response in
+        pigeonApi.flutterEchoString(pigeonInstance: pigeonInstance, aString: aString) {
+          response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -1862,7 +1953,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         aMap: [String?: ProxyApiTestClass?],
         completion: @escaping (Result<[String?: ProxyApiTestClass?], Error>) -> Void
       ) {
-        pigeonApi.flutterEchoProxyApiMap(pigeonInstance: pigeonInstance, aMap: aMap) { response in
+        pigeonApi.flutterEchoProxyApiMap(pigeonInstance: pigeonInstance, aMap: aMap) {
+          response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -1874,7 +1966,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
 
       func callFlutterEchoEnum(
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
-        anEnum: ProxyApiTestEnum, completion: @escaping (Result<ProxyApiTestEnum, Error>) -> Void
+        anEnum: ProxyApiTestEnum,
+        completion: @escaping (Result<ProxyApiTestEnum, Error>) -> Void
       ) {
         pigeonApi.flutterEchoEnum(pigeonInstance: pigeonInstance, anEnum: anEnum) { response in
           switch response {
@@ -1921,7 +2014,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, anInt: Int64?,
         completion: @escaping (Result<Int64?, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoNullableInt(pigeonInstance: pigeonInstance, anInt: anInt) { response in
+        pigeonApi.flutterEchoNullableInt(pigeonInstance: pigeonInstance, anInt: anInt) {
+          response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -1932,7 +2026,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoNullableDouble(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aDouble: Double?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aDouble: Double?,
         completion: @escaping (Result<Double?, Error>) -> Void
       ) {
         pigeonApi.flutterEchoNullableDouble(pigeonInstance: pigeonInstance, aDouble: aDouble) {
@@ -1947,7 +2042,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoNullableString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String?,
         completion: @escaping (Result<String?, Error>) -> Void
       ) {
         pigeonApi.flutterEchoNullableString(pigeonInstance: pigeonInstance, aString: aString) {
@@ -1966,7 +2062,9 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         aUint8List: FlutterStandardTypedData?,
         completion: @escaping (Result<FlutterStandardTypedData?, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoNullableUint8List(pigeonInstance: pigeonInstance, aList: aUint8List) {
+        pigeonApi.flutterEchoNullableUint8List(
+          pigeonInstance: pigeonInstance, aList: aUint8List
+        ) {
           response in
           switch response {
           case .success(let res):
@@ -1978,7 +2076,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoNullableList(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aList: [Any?]?,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aList: [Any?]?,
         completion: @escaping (Result<[Any?]?, Error>) -> Void
       ) {
         pigeonApi.flutterEchoNullableList(pigeonInstance: pigeonInstance, aList: aList) {
@@ -1996,7 +2095,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
         aMap: [String?: Any?]?, completion: @escaping (Result<[String?: Any?]?, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoNullableMap(pigeonInstance: pigeonInstance, aMap: aMap) { response in
+        pigeonApi.flutterEchoNullableMap(pigeonInstance: pigeonInstance, aMap: aMap) {
+          response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -2008,7 +2108,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
 
       func callFlutterEchoNullableEnum(
         pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
-        anEnum: ProxyApiTestEnum?, completion: @escaping (Result<ProxyApiTestEnum?, Error>) -> Void
+        anEnum: ProxyApiTestEnum?,
+        completion: @escaping (Result<ProxyApiTestEnum?, Error>) -> Void
       ) {
         pigeonApi.flutterEchoNullableEnum(pigeonInstance: pigeonInstance, anEnum: anEnum) {
           response in
@@ -2026,8 +2127,9 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         aProxyApi: ProxyApiSuperClass?,
         completion: @escaping (Result<ProxyApiSuperClass?, Error>) -> Void
       ) {
-        pigeonApi.flutterEchoNullableProxyApi(pigeonInstance: pigeonInstance, aProxyApi: aProxyApi)
-        { response in
+        pigeonApi.flutterEchoNullableProxyApi(
+          pigeonInstance: pigeonInstance, aProxyApi: aProxyApi
+        ) { response in
           switch response {
           case .success(let res):
             completion(.success(res))
@@ -2052,7 +2154,8 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
       }
 
       func callFlutterEchoAsyncString(
-        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass, aString: String,
+        pigeonApi: PigeonApiProxyApiTestClass, pigeonInstance: ProxyApiTestClass,
+        aString: String,
         completion: @escaping (Result<String, Error>) -> Void
       ) {
         pigeonApi.flutterEchoAsyncString(pigeonInstance: pigeonInstance, aString: aString) {
@@ -2081,7 +2184,9 @@ class ProxyApiDelegate: ProxyApiTestsPigeonProxyApiDelegate {
         return ProxyApiSuperClass()
       }
 
-      func aSuperMethod(pigeonApi: PigeonApiProxyApiSuperClass, pigeonInstance: ProxyApiSuperClass)
+      func aSuperMethod(
+        pigeonApi: PigeonApiProxyApiSuperClass, pigeonInstance: ProxyApiSuperClass
+      )
         throws
       {}
     }