+++ packages/pigeon/test/dart_generator_test.dart
@@ -1711,17 +1711,22 @@ name: foobar
   });
 
   test('connection error contains channel name', () {
-    final Root root = Root(apis: <Api>[
-      AstHostApi(name: 'Api', methods: <Method>[
-        Method(
-          name: 'method',
-          location: ApiLocation.host,
-          parameters: <Parameter>[],
-          returnType:
-              const TypeDeclaration(baseName: 'Output', isNullable: false),
-        )
-      ])
-    ], classes: <Class>[], enums: <Enum>[]);
+    final Root root = Root(
+      apis: <Api>[
+        AstHostApi(name: 'Api', methods: <Method>[
+          Method(
+            name: 'method',
+            location: ApiLocation.host,
+            parameters: <Parameter>[],
+            returnType:
+                const TypeDeclaration(baseName: 'Output', isNullable: false),
+          )
+        ])
+      ],
+      classes: <Class>[],
+      enums: <Enum>[],
+      containsHostApi: true,
+    );
     final StringBuffer sink = StringBuffer();
     const DartGenerator generator = DartGenerator();
     generator.generate(